make reader an internal module and refactor the parsing of binary files
